requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
httpx>=0.24.0
akshare>=1.12.0
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=14.0.0
apscheduler>=3.10.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
pyyaml>=6.0
python-dotenv>=1.0.0
jinja2>=3.0.0
requests>=2.31.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
sqlalchemy>=2.0.0
pymysql>=1.1.0
polars>=0.20.0
duckdb>=0.10.0
redis>=4.5.0
kafka-python>=2.0.2
tqdm>=4.65.0
psutil>=5.9.0
loguru>=0.7.0

# 性能优化相关依赖
# =============================================================================
aiohttp>=3.9.0
aiosignal>=1.3.0
frozenlist>=1.4.0
cachetools>=5.3.0
prometheus-client>=0.19.0

# DataHub 集成 (元数据管理)
# =============================================================================
# 基础 CLI 和常用插件 (Doris, Kafka, Great Expectations)
# acryl-datahub[doris,kafka,ge]>=0.12.0
# 或使用以下命令安装:
# pip install 'acryl-datahub[doris,kafka,ge]'
